+++ ticketing/templates/ticketing/ticket_items.tmpl
@@ -43,6 +43,11 @@ Act Fees
     </div>
     <div id="collapseOne" class="collapse" aria-labelledby="headingOne" data-parent="#accordion">
       <div class="card-body">
+<h4>Preview</h4>
+<div class="panel panel-default">
+  <br>
+{% include "form_table.tmpl" with form=act_pay_form %}
+</div>
 {% include "ticketing/ticket_list_display.tmpl" with events=act_fees%}
      </div>
     </div>
@@ -61,6 +66,11 @@ Vendor Fees
     </div>
     <div id="collapseTwo" class="collapse" aria-labelledby="headingTwo" data-parent="#accordion">
       <div class="card-body">
+<h4>Preview</h4>
+<div class="panel panel-default">
+  <br>
+{% include "form_table.tmpl" with form=vendor_pay_form %}
+</div>
 {% include "ticketing/ticket_list_display.tmpl" with events=vendor_fees%}
       </div>
     </div>
